algophant-backend/AlgoPhantAPI/requirements.txt
asgiref==3.5.2
cffi==1.15.1
dj-database-url==1.0.0
Django==4.1.2
django-cors-headers==3.13.0
django-heroku==0.3.1
djangorestframework==3.14.0
gunicorn==20.1.0
msgpack==1.0.4
psycopg2==2.9.4
py-algorand-sdk==1.18.0
pycparser==2.21
pycryptodomex==3.15.0
PyNaCl==1.5.0
pytz==2022.4
sqlparse==0.4.3
whitenoise==6.2.0
